convert commands to own files
