+++ commands/ping.js
@@ -1,0 +1,8 @@
+module.exports = {
+  name: 'ping',
+  description: 'Ping the bot',
+  aliases: [],
+  command: (message, arguments, serverQueue, client) => {
+    return message.channel.send('pong');
+  },
+};
