+++ commands/stop.js
@@ -1,0 +1,19 @@
+module.exports = {
+  name: 'stop',
+  description: 'Stop playback',
+  aliases: [],
+  command: (message, arguments, serverQueue, client) => {
+    if (!message.member.voice.channel) {
+      return message.channel.send(
+        'You have to be in a voice channel to stop the music!'
+      );
+    }
+
+    if (!serverQueue) {
+      return message.channel.send('There is no song that I could stop!');
+    }
+
+    serverQueue.songs = [];
+    serverQueue.audioPlayer.stop();
+  },
+};
